// src/api/schemas/getDashboard200DataComisionesItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAC API
 * Api para uso interno WAC.
 * OpenAPI spec version: 1.0.16
 */
import type { GetDashboard200DataComisionesItemMesesItem } from './getDashboard200DataComisionesItemMesesItem';

export type GetDashboard200DataComisionesItem = {
  ano?: string;
  meses?: GetDashboard200DataComisionesItemMesesItem[];
};