// src/api/schemas/getDashboard200DataKpisVentas.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAC API
 * Api para uso interno WAC.
 * OpenAPI spec version: 1.0.16
 */
import type { GetDashboard200DataKpisVentasMes } from './getDashboard200DataKpisVentasMes';
import type { GetDashboard200DataKpisVentasMesAnterior } from './getDashboard200DataKpisVentasMesAnterior';
import type { GetDashboard200DataKpisVentasAno } from './getDashboard200DataKpisVentasAno';
import type { GetDashboard200DataKpisVentasAnoAnterior } from './getDashboard200DataKpisVentasAnoAnterior';

export type GetDashboard200DataKpisVentas = {
  mes?: GetDashboard200DataKpisVentasMes;
  mes_anterior?: GetDashboard200DataKpisVentasMesAnterior;
  ano?: GetDashboard200DataKpisVentasAno;
  ano_anterior?: GetDashboard200DataKpisVentasAnoAnterior;
};